requirements.txt
flask==3.0.3
google-generativeai==0.7.2
rank-bm25==0.2.2
numpy>=2,<3
python-dotenv==1.0.1
pydantic==2.9.2
